Feature: Add support for custom config fragments
Previously the configuration was generated from a big monolithic template
and allowed custom options to be specified via the options hash. This does
not work well for profiles wanting to change some ssh settings without
managing the whole ssh configuration.

This commit switches sshd_config to be generated via concat, making the
previous monolithic template the first fragment. Apart from this it
adds a ssh::server::custom_config type which allows adding config blocks
at the end of the configuration.