Add simple tests for ssh_keygen function
To make the parser function testable allow basedir being
passed as argument instead of using the hardcoded path.
Use this new feature to pass a tempdir as path and test
against those.
While at it I fixed a bug in the code which assumed that an
array is a string.