Allow management of host certificate and known_hosts
Now that we have a method for signing a host certificate,
make it possible to actually let the ssh::server class configure
do that for the hostkey and configure it as HostCertificate.
While at it, add a new parameter trusted_ssh_certificates that
configures arbitrary ca certificates in global known_hosts.
Added an acceptance test for both.