Fix argument checking
First, we have three required arguments now. Second, relying
on what I considered tricky in the case when statement is well ...
unreliable. At least in ruby 1.8 case when is somewhat racy.
So let's explicitly check for kind_of? which is the intension of
the code. As a nice side effect it's more clear now.